backend/requirements.txt
asgiref==3.4.1
click==8.0.3
fastapi==0.70.0
h11==0.12.0
pydantic==1.8.2
starlette==0.16.0
typing-extensions==3.10.0.2
uvicorn==0.14.0
mysql-connector-python==8.0.27
PyJWT==1.7.1
Werkzeug==2.0.2
fastapi-mail==1.0.2
emails==0.6
GitPython==3.1.24
gitdb==4.0.9
requests==2.26.0
python-multipart==0.0.5
boto3==1.21.2
aiomysql==0.0.22
cryptography==36.0.2
psutil==5.9.0